pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nightwatch"
version = "0.1.0"
description = "Open-source epilepsy monitoring system for Raspberry Pi"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Nightwatch Contributors" }
]
keywords = ["epilepsy", "monitoring", "raspberry-pi", "health", "seizure"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core
    "pyzmq>=25.0",
    "msgpack>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "watchdog>=3.0",

    # Detectors
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "scipy>=1.10",

    # Dashboard
    "fastapi>=0.100",
    "uvicorn[standard]>=0.22",
    "websockets>=11.0",
    "httpx>=0.27",
    "aiohttp>=3.8",
    "aiosqlite>=0.19",
    "jinja2>=3.1",
]

[project.optional-dependencies]
pi = [
    "spidev>=3.6",
    "RPi.GPIO>=0.7",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "respx>=0.20",
    "black>=23.0",
    "ruff>=0.0.280",
    "mypy>=1.4",
]

[project.scripts]
nightwatch = "nightwatch.__main__:main"

[project.urls]
Homepage = "https://github.com/nightwatch/nightwatch"
Documentation = "https://github.com/nightwatch/nightwatch#readme"
Repository = "https://github.com/nightwatch/nightwatch"

[tool.hatch.build.targets.wheel]
packages = ["nightwatch"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
